move get_api_brain_definition_as_json_schema.py
